fix(i9w): fix accidental UTC output
The way libtai is currently used generates beats in UTC
that include the leap seconds.
We need to hack around the normal libtai API a bit
to obtain the raw fractional seconds we can convert to beatTAI.

This fix matches the output of the Go implementation.
I have checked it, and it's 37 seconds ahead of UTC,
which is correct for TAI since 2017.

Update the prefix from "@" to ":".